Reference Particle: GPU Capable
Make all methods of the reference particle type:
- GPU capable (`__host__ __device__`)
- forced inline

Create a separate header file for the type.
